requirements.txt
torch
torchvision
opencv-python
pillow
nibabel
customtkinter
numpy
scikit-image
reportlab